+++ settings.gradle
@@ -4,5 +4,5 @@ include 'wss-web'
 include 'wss-persistance'
 include 'wss-service'
 include 'wss'
-
+include 'wss-auth'
