+++ wss-auth/src/main/java/club/csmrobotics/wss/security/InvalidJwtAuthenticationException.java
@@ -1,0 +1,9 @@
+package club.csmrobotics.wss.security;
+
+import org.springframework.security.core.AuthenticationException;
+
+public class InvalidJwtAuthenticationException extends AuthenticationException {
+    public InvalidJwtAuthenticationException(String e) {
+        super(e);
+    }
+}
