+++ wss-auth/src/main/java/club/csmrobotics/wss/security/domain/UserDTO.java
@@ -1,0 +1,16 @@
+package club.csmrobotics.wss.security.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserDTO implements Serializable {
+    private String username;
+    private String password;
+    private String roles;
+}
