+++ wss-auth/src/main/java/club/csmrobotics/wss/security/service/UserDetailsServiceImpl.java
@@ -1,0 +1,23 @@
+package club.csmrobotics.wss.security.service;//package club.csmrobotics.wss.security;
+
+import club.csmrobotics.wss.security.dao.UserRep;
+import club.csmrobotics.wss.security.domain.UserPrincipal;
+import lombok.AllArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Component;
+
+import java.text.MessageFormat;
+
+@Component
+@AllArgsConstructor
+public class UserDetailsServiceImpl implements UserDetailsService {
+    private final UserRep userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return new UserPrincipal(userRepository.findByUsername(username)
+                .orElseThrow(() -> new UsernameNotFoundException(MessageFormat.format("Username: {0} not found", username))));
+    }
+}
