+++ wss-auth/src/main/resources/jwtExample.conf
