+++ wss-domain/build.gradle
@@ -1,5 +1,3 @@
-group 'rclub'
-version '1.0-SNAPSHOT'
 
 
 useCommons()
