+++ wss-persistance/build.gradle
@@ -1,5 +1,4 @@
-group 'rclub'
-version '1.0-SNAPSHOT'
+
 
 useCommons()
 useSpring()
