--- wss-persistance/src/main/java/rclub/wss/persistance/auth/UserRep.java
@@ -1,10 +1,0 @@
-package rclub.wss.persistance.auth;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import rclub.wss.domain.auth.User;
-
-import java.util.Optional;
-
-public interface UserRep extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-}
