--- wss-persistance/src/main/java/rclub/wss/persistance/data/PackageRep.java
@@ -1,9 +1,0 @@
-package rclub.wss.persistance.data;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-import rclub.wss.domain.data.Package;
-
-@Repository
-public interface PackageRep extends CrudRepository<Package, Long> {
-}
