+++ wss-service/src/main/java/club/csmrobotics/wss/service/PackageNotFoundException.java
@@ -1,0 +1,9 @@
+package club.csmrobotics.wss.service;
+
+import java.text.MessageFormat;
+
+public class PackageNotFoundException extends RuntimeException {
+    public PackageNotFoundException(Long id) {
+        super(MessageFormat.format("Package: {0} not found", id));
+    }
+}
