--- wss-service/src/main/java/rclub/wss/service/PackageNotFoundException.java
@@ -1,9 +1,0 @@
-package rclub.wss.service;
-
-import java.text.MessageFormat;
-
-public class PackageNotFoundException extends RuntimeException {
-    public PackageNotFoundException(Long id) {
-        super(MessageFormat.format("Package: {0} not found", id));
-    }
-}
