--- wss-web/src/main/java/rclub/wss/security/InvalidJwtAuthenticationException.java
@@ -1,9 +1,0 @@
-package rclub.wss.security;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class InvalidJwtAuthenticationException extends AuthenticationException {
-    public InvalidJwtAuthenticationException(String e) {
-        super(e);
-    }
-}
