--- wss-web/src/main/java/rclub/wss/security/JwtAuthorizationFilter.java
@@ -1,56 +1,0 @@
-package rclub.wss.security;
-
-import com.auth0.jwt.JWT;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
-import rclub.wss.persistance.auth.UserRep;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-import static com.auth0.jwt.algorithms.Algorithm.HMAC512;
-
-public class JwtAuthorizationFilter extends BasicAuthenticationFilter {
-    private UserRep userRepository;
-
-    public JwtAuthorizationFilter(AuthenticationManager authenticationManager, UserRep userRepository) {
-        super(authenticationManager);
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain) throws IOException, ServletException {
-        String header = request.getHeader(JwtProperties.HEADER_STRING);
-
-        if (header == null || !header.startsWith(JwtProperties.TOKEN_PREFIX)) {
-            chain.doFilter(request, response);
-            return;
-        }
-
-        // If header is present, try grab user principal from database and perform authorization
-        Authentication authentication = getUsernamePasswordAuthentication(request);
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-
-        // Continue filter execution
-        chain.doFilter(request, response);
-    }
-
-    private Authentication getUsernamePasswordAuthentication(HttpServletRequest request) {
-        String token = request.getHeader(JwtProperties.HEADER_STRING).replace(JwtProperties.TOKEN_PREFIX, "");
-        String userName = JWT.require(HMAC512(JwtProperties.SECRET.getBytes()))
-                .build()
-                .verify(token)
-                .getSubject();
-
-        if (userName == null) return null;
-        return userRepository.findByUsername(userName).map(value ->
-                new UsernamePasswordAuthenticationToken(userName, null, new UserPrincipal(value).getAuthorities())
-        ).orElse(null);
-    }
-}
