--- wss-web/src/main/java/rclub/wss/security/UserDetailsServiceImpl.java
@@ -1,23 +1,0 @@
-package rclub.wss.security;
-
-import lombok.AllArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-import rclub.wss.persistance.auth.UserRep;
-
-import java.text.MessageFormat;
-
-@Component
-@AllArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private UserRep userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return new UserPrincipal(userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException(MessageFormat.format("Username: {0} not found", username))));
-    }
-}
