--- wss-web/src/main/java/rclub/wss/web/RestExceptionHandler.java
@@ -1,29 +1,0 @@
-package rclub.wss.web;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.context.request.WebRequest;
-import rclub.wss.security.InvalidJwtAuthenticationException;
-import rclub.wss.service.PackageNotFoundException;
-
-import static org.springframework.http.HttpStatus.UNAUTHORIZED;
-import static org.springframework.http.ResponseEntity.notFound;
-import static org.springframework.http.ResponseEntity.status;
-
-@RestControllerAdvice
-@Slf4j
-public class RestExceptionHandler {
-
-    @ExceptionHandler(value = {PackageNotFoundException.class})
-    public ResponseEntity<?> vehicleNotFound(PackageNotFoundException ex, WebRequest request) {
-        return notFound().build();
-    }
-
-    @ExceptionHandler(value = {InvalidJwtAuthenticationException.class})
-    public ResponseEntity<?> invalidJwtAuthentication(InvalidJwtAuthenticationException ex, WebRequest request) {
-        log.debug("handling InvalidJwtAuthenticationException from" + request.toString());
-        return status(UNAUTHORIZED).build();
-    }
-}
