+++ wss/build.gradle
@@ -1,6 +1,3 @@
-group 'rclub'
-version '1.0-SNAPSHOT'
-
 
 useCommons()
 useSpring()
